fix mobile merge conflicts
