+++ src/components/Tabs/TabsUI.tsx
@@ -3,9 +3,19 @@ import isEqual from 'lodash/isEqual';
 import classNames from 'classnames';
 import { ChevronLeft, ChevronRight } from 'wix-ui-icons-common';
 import { ALIGNMENT, VARIANT } from './constants';
+import { KEY_CODES } from '../../common/constants';
 import { Tab } from './Tab';
 import style from './Tabs.st.css';
-import { isSelectKey } from './Tabs';
+
+export function isSelectKey(keyCode: number) {
+  switch (keyCode) {
+    case KEY_CODES.ENTER:
+    case KEY_CODES.SPACEBAR:
+      return true;
+    default:
+      return false;
+  }
+}
 
 export interface TabItem {
   /** Title of the tab */
